+++ src/runtime/cgo_mmap.go
@@ -25,7 +25,7 @@ var _cgo_munmap unsafe.Pointer
 // (used by sysAlloc) is called in a lot of low-level parts of the runtime and
 // callers often assume it won't acquire any locks.
 //go:nosplit
-func mmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uint32) (unsafe.Pointer, int) {
+func mmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uintptr) (unsafe.Pointer, int) {
 	if _cgo_mmap != nil {
 		// Make ret a uintptr so that writing to it in the
 		// function literal does not trigger a write barrier.
@@ -53,11 +53,11 @@ func munmap(addr unsafe.Pointer, n uintptr) {
 }
 
 // sysMmap calls the mmap system call. It is implemented in assembly.
-func sysMmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uint32) (p unsafe.Pointer, err int)
+func sysMmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uintptr) (p unsafe.Pointer, err int)
 
 // callCgoMmap calls the mmap function in the runtime/cgo package
 // using the GCC calling convention. It is implemented in assembly.
-func callCgoMmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uint32) uintptr
+func callCgoMmap(addr unsafe.Pointer, n uintptr, prot, flags, fd int32, off uintptr) uintptr
 
 // sysMunmap calls the munmap system call. It is implemented in assembly.
 func sysMunmap(addr unsafe.Pointer, n uintptr)
